+++ frontend/src/models/Window/repository.ts
@@ -1,5 +1,4 @@
-import { API_URL } from '@/config/config';
-import { updateOnlineStatus } from '@/models/User/repository';
+import { executeLogout } from '@/models/User/auth';
 
 const IDLE_TIMEOUT = 10000; // 10秒
 
@@ -21,15 +20,7 @@ export const resetTimer = () => {
 
   idleTimer = window.setTimeout(() => {
     try {
-      updateOnlineStatus(false);
-      fetch(`${API_URL}/api/logout/`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        credentials: 'include',
-      });
-      clearUserSession();
+      executeLogout();
       window.location.href = '/login';
     } catch (error) {
       console.error('Logout API call failed:', error);
